refactor files and unit tests
