+++ sample_size/sample_size_calculator.py
@@ -1,12 +1,13 @@
+from typing import Any
 from typing import List
 
 import numpy as np
 from statsmodels.stats.power import NormalIndPower
 from statsmodels.stats.power import TTestIndPower
 
-from sample_size.sample_size_calculator.variables import Boolean
-from sample_size.sample_size_calculator.variables import Numeric
-from sample_size.sample_size_calculator.variables import Ratio
+from sample_size.variables import Boolean
+from sample_size.variables import Numeric
+from sample_size.variables import Ratio
 
 DEFAULT_ALPHA = 0.05
 DEFAULT_POWER = 0.8
@@ -22,11 +23,11 @@ class PowerAnalysisParameters:
     """
 
     def __init__(
-            self,
-            metric_variance: float,
-            mde: float,
-            alpha: float = DEFAULT_ALPHA,
-            power: float = DEFAULT_POWER,
+        self,
+        metric_variance: float,
+        mde: float,
+        alpha: float = DEFAULT_ALPHA,
+        power: float = DEFAULT_POWER,
     ):
         self.metric_variance = metric_variance
         self.mde = mde
@@ -50,31 +51,31 @@ class SampleSizeCalculator:
         self.numeric_metrics: List[PowerAnalysisParameters] = []
         self.ratio_metrics: List[PowerAnalysisParameters] = []
 
-    def register_bool_metric(self, mde: float, probability: float):
+    def register_bool_metric(self, mde: float, probability: float) -> None:
         variable = Boolean(probability)
         metric = PowerAnalysisParameters(variable.variance, mde, self.alpha, self.power)
         self.boolean_metrics.append(metric)
 
-    def register_numeric_metric(self, mde: float, variance: float):
+    def register_numeric_metric(self, mde: float, variance: float) -> None:
         variable = Numeric(variance)
         metric = PowerAnalysisParameters(variable.variance, mde, self.alpha, self.power)
         self.numeric_metrics.append(metric)
 
     def register_ratio_metric(
-            self,
-            mde: float,
-            numerator_mean: float,
-            numerator_variance: float,
-            denominator_mean: float,
-            denominator_variance: float,
-            covariance: float,
-    ):
+        self,
+        mde: float,
+        numerator_mean: float,
+        numerator_variance: float,
+        denominator_mean: float,
+        denominator_variance: float,
+        covariance: float,
+    ) -> None:
         variable = Ratio(numerator_mean, numerator_variance, denominator_mean, denominator_variance, covariance)
         metric = PowerAnalysisParameters(variable.variance, mde, self.alpha, self.power)
         self.ratio_metrics.append(metric)
 
     @staticmethod
-    def _get_single_sample_size(metric, power_analysis_type):
+    def _get_single_sample_size(metric: PowerAnalysisParameters, power_analysis_type: Any) -> float:
         effect_size = metric.mde / float(np.sqrt(metric.metric_variance))
         power_analysis = power_analysis_type()
         sample_size = int(
@@ -88,14 +89,18 @@ class SampleSizeCalculator:
         )
         return sample_size
 
-    def get_overall_sample_size(self):
-        # Supports the sample size calculation for single metric now
+    def get_overall_sample_size(self) -> float:
+        # Supports the sample size calculation for single metric now.
+        # The current structure is set up to support multiple metrics in the future.
+        sample_size = float("nan")
         if self.boolean_metrics:
             for metric in self.boolean_metrics:
-                return self._get_single_sample_size(metric, NormalIndPower)
+                sample_size = self._get_single_sample_size(metric, NormalIndPower)
         if self.numeric_metrics:
             for metric in self.numeric_metrics:
-                return self._get_single_sample_size(metric, TTestIndPower)
+                sample_size = self._get_single_sample_size(metric, TTestIndPower)
         if self.ratio_metrics:
             for metric in self.ratio_metrics:
-                return self._get_single_sample_size(metric, TTestIndPower)
+                sample_size = self._get_single_sample_size(metric, TTestIndPower)
+
+        return sample_size
